Add files via upload
The task for this project was to determine the floor installation cost for specific rooms. We were tasked to create a different python file for each type of room, prompt the user for information regarding the room, and create formulas to deliver the cost of the installation for each type of room.